rename: 공통 컴포넌트 폴더이름 변경
