+++ src/app/(home)/_components/FilterSection.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { useQueryBuilder } from "@/hooks/useUrlParams";
-import { FilterDropDown } from "@/components/filter/FilterDropDown";
+import { FilterDropDown } from "@/components/filters/FilterDropDown";
 import { DIRECTION_OPTIONS, LOCATION_OPTIONS, SORT_OPTIONS } from "@/constants/filter";
 
 export default function FilterSection() {
